backend/requirements.txt
fastapi
kafka
pymongo
json
os
python-dotenv
logging
numpy
random
scipy
threading